requirements.txt
absl-py==1.1.0
alabaster==0.7.12
anyio==3.6.1
asgiref==3.5.2
astor==0.8.1
astunparse==1.6.3
awscli==1.25.1
Babel==2.9.1
bce-python-sdk==0.8.64
botocore==1.27.1
cachetools==5.2.0
certifi==2022.5.18.1
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coverage==6.4.1
cssselect==1.1.0
cssutils==2.3.0
cycler==0.11.0
Cython==0.29.26
decorator==5.1.0
distlib==0.3.4
docutils==0.16
et-xmlfile==1.1.0
fastapi==0.78.0
filelock==3.4.2
flake8==4.0.1
Flask-Babel==2.0.0
flatbuffers==1.12
fonttools==4.28.5
future==0.18.2
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h11==0.13.0
h5py==3.7.0
identify==2.4.1
idna==3.3
imageio==2.13.5
imagesize==1.3.0
imgaug==0.4.0
importlib-metadata==4.11.4
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==1.0.0
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==14.0.1
lmdb==1.3.0
lxml==4.7.1
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
networkx==2.6.3
nodeenv==1.6.0
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.6.0.66
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
Pillow==9.1.1
platformdirs==2.4.1
pre-commit==2.16.0
premailer==3.10.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post2
pycodestyle==2.8.0
pycryptodome==3.12.0
pydantic==1.9.1
pyflakes==2.4.0
Pygments==2.12.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.20.0
python-Levenshtein==0.12.2
pytz==2022.1
PyWavelets==1.2.0
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.6.0
scikit-image==0.19.1
scikit-learn==1.1.1
scipy==1.7.3
Shapely==1.8.0
shellcheck-py==0.8.0.3
six==1.16.0
sklearn==0.0
sniffio==1.2.0
snowballstemmer==2.2.0
Sphinx==5.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
# Editable install with no version control (src==0.1.0)
-e /home/hoola/code/mobilenet_parkinson
starlette==0.19.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2021.11.2
toml==0.10.2
tqdm==4.62.3
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
virtualenv==20.13.0
visualdl==2.2.2
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
